elevation/requirements.txt
GDAL>=3.6
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
